requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-ckeditor==5.0.3
django-modeltranslation==0.11
django-reversetag==0.3.2
django-taggit==0.20.2
django-treebeard==4.0.1
django-wysiwyg-redactor==0.4.9
easy-thumbnails==2.3
oauthlib==1.1.2
Pillow==3.2.0
PyJWT==1.4.1
pysolr==3.5.0
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
Whoosh==2.7.4
